containers/backend/requirements.txt
Django==3.0.6
django-cors-middleware==1.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
djoser==2.0.3
envparse==0.2.0
# tests
factory-boy==2.12.0
